udpate parsing

+++ Project1/4th/GPSParsing/gps_parsing.py
@@ -119,8 +119,8 @@ def get_data(db_path, table_name, column_name):
 #                             print(lon_dd, lon_mm)
 #                             print(f'lon : {lon_final} / lat : {lat_final}')
 
-                            lon_dd = lon/10000000
-                            lat_dd = lat/10000000
+                            lon_dd = int(lon/10000000)
+                            lat_dd = int(lat/10000000)
 #                             lon_mm = (lon%10000000)/10000000
 #                             lat_mm = (lat%10000000)/10000000
                             lon_mm = ((lon/10000000-lon_dd)*100)/60
